cleanup and consolidate graphql types into one file
