+++ src/schema/v2/order/index.ts
@@ -1,2 +1,3 @@
 export { MeOrder } from "schema/v2/order/MeOrder"
 export { updateOrderMutation } from "schema/v2/order/updateOrderMutation"
+export { setOrderFulfillmentOptionMutation } from "schema/v2/order/setOrderFulfillmentOptionMutation"
